nextjs-dashboard/api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
python-dotenv
psycopg2-binary  
